Prepare website previews (#3)
For theia-website, we switch from legacy gh-pages to the workflow based
approach.

To deploy previews for theia-website, we create a dedicated repository
called theia-website-previews.

In this repository, a workflow will create a directory for each PR that
is opened against theia-website. When the PR is closed, the respective
directory will be removed again via workflow.

Signed-off-by: Olaf Lessenich <olessenich@eclipsesource.com>